add user endpoints

+++ src/activity/activity.service.ts
@@ -1,6 +1,7 @@
-import { User } from '@prisma/client';
-import { Injectable } from '@nestjs/common';
+import { Injectable, UnauthorizedException } from '@nestjs/common';
+import { Activity, Role, UsersOnProjects } from '@prisma/client';
 import { PrismaService } from 'src/prisma/prisma.service';
+import { TrackingRequest } from '../auth/middleware/auth.middleware';
 import { CreateActivityDto, UpdateActivityDto } from './dto';
 
 @Injectable()
@@ -13,18 +14,27 @@ export class ActivityService {
    * @param dto
    * @returns activity Activity
    */
-  async create(dto: CreateActivityDto) {
-    const activity = await this.prisma.activity.create({
-      data: {
-        ...dto,
-      },
-      include: {
-        user: true,
-        activityType: true,
-        project: true,
-      },
+  async create(req: TrackingRequest, dto: CreateActivityDto): Promise<Activity> {
+    const projectUser: UsersOnProjects = await this.prisma.usersOnProjects.findUnique({
+      where: { userId: req.userId, projectId: dto.projectId },
+      rejectOnNotFound: true,
     });
-    return activity;
+
+    if (projectUser != null) {
+      const activity = await this.prisma.activity.create({
+        data: {
+          ...dto,
+        },
+        include: {
+          user: true,
+          activityType: true,
+          project: true,
+        },
+      });
+      return activity;
+    }
+
+    throw new UnauthorizedException();
   }
 
   /**
@@ -34,21 +44,25 @@ export class ActivityService {
    * @param dto
    * @returns activity Activity
    */
-  async update(activityId: number, dto: UpdateActivityDto) {
-    const activity = await this.prisma.activity.update({
-      where: {
-        id: activityId,
-      },
-      data: {
-        ...dto,
-      },
-      include: {
-        user: true,
-        activityType: true,
-        project: true,
-      },
-    });
-    return activity;
+  async update(req: TrackingRequest, activityId: number, dto: UpdateActivityDto): Promise<Activity> {
+    if (this.canEditActivity(activityId, req.userId, req.userRole)) {
+      const activity = await this.prisma.activity.update({
+        where: {
+          id: activityId,
+        },
+        data: {
+          ...dto,
+        },
+        include: {
+          user: true,
+          activityType: true,
+          project: true,
+        },
+      });
+      return activity;
+    }
+
+    throw new UnauthorizedException();
   }
 
   /**
@@ -56,12 +70,16 @@ export class ActivityService {
    *
    * @param acitivityId
    */
-  async delete(acitivityId: number) {
-    await this.prisma.activity.delete({
-      where: {
-        id: acitivityId,
-      },
-    });
+  async delete(req: TrackingRequest, activityId: number): Promise<void> {
+    if (this.canEditActivity(activityId, req.userId, req.userRole)) {
+      await this.prisma.activity.delete({
+        where: {
+          id: activityId,
+        },
+      });
+    }
+
+    throw new UnauthorizedException();
   }
 
   /**
@@ -69,20 +87,57 @@ export class ActivityService {
    *
    * @param activityId
    */
-  async get(activityId: number) {
-    return this.prisma.activity.findFirst({
+  async get(req: TrackingRequest, activityId: number) {
+    const activity: Activity = await this.prisma.activity.findFirst({
       where: {
         id: activityId,
       },
     });
+
+    if (req.userRole === Role.ADMIN || activity.userId === req.userId) {
+      return activity;
+    }
+
+    const projectUser: UsersOnProjects = await this.prisma.usersOnProjects.findUnique({
+      where: { userId: req.userId, projectId: activity.projectId },
+      rejectOnNotFound: true,
+    });
+    if (req.userRole === Role.SCIENTIST && projectUser != null) {
+      return activity;
+    }
+
+    throw new UnauthorizedException();
   }
 
   /**
-   * Get All Activities
-   *
-   * @returns PrismaPromise<...>
+   * Get All Activities of a project
    */
-  async getAll() {
-    return this.prisma.activity.findMany();
+  async getProjectActivities(req: TrackingRequest, projectId: number): Promise<Activity[]> {
+    const projectUser: UsersOnProjects = await this.prisma.usersOnProjects.findUnique({
+      where: { userId: req.userId, projectId },
+      rejectOnNotFound: false,
+    });
+
+    if (req.userRole === Role.ADMIN || (req.userRole === Role.SCIENTIST && projectUser != null)) {
+      return this.prisma.activity.findMany({ where: { projectId } });
+    }
+
+    throw new UnauthorizedException();
+  }
+
+  /**
+   * Returns if a user can edit an acitivity
+   */
+  private async canEditActivity(acitivityId: number, userId: number, userRole: Role): Promise<boolean> {
+    if (userRole === Role.ADMIN) {
+      return true;
+    }
+
+    const activity: Activity = await this.prisma.activity.findFirst({ where: { id: acitivityId } });
+    if (activity.userId === userId) {
+      return true;
+    }
+
+    return false;
   }
 }
